Terminated the project

+++ Image/sobel_filter.h
@@ -2,11 +2,11 @@
 #define SOBEL_FILTER_H
 
 #include <stdlib.h>
-#include <SDL.h>
+#include <SDL/SDL.h>
 #include <math.h>
 
 int gradient(int x, int y);
-int updatePixel2(Uint8 pixels[]);
+int updatePixel2(int pixels[]);
 void Sfilter(SDL_Surface* image, SDL_Surface* n_image);
 
 #endif
